+++ packages/appcd/src/actions/pm-install.js
@@ -1,9 +1,0 @@
-export default async ({ console }) => {
-	const { action } = await import('../commands/pm/install');
-	await action({
-		argv: {
-			plugins: [ 'default' ]
-		},
-		console
-	});
-};